api/requirements.txt
Django==3.0.4
django-cors-headers==3.2.1
django-environ==0.4.5
django-extensions==2.2.8
django-filter==2.2.0
django-spa==0.3.5
djangorestframework==3.11.0
drf-extensions=0.6.0
